Syntax highlight LaTeX and PDF outputs
Requires pygments installed with, e.g.
$ sudo apt-get install python-pygments
$ sudo gem install pygments.rb

- Move highlight method from htmlutils.rb to textutils.rb and stop
  unescaping inside the method
- Call highlight from list_body method in latexbuilder.rb
- Add preamble needed to review.tex.erb
