+++ lib/review/htmlutils.rb
@@ -31,29 +31,5 @@ module ReVIEW
     def strip_html(str)
       str.gsub(/<\/?[^>]*>/, "")
     end
-
-    def highlight(ops)
-      body = ops[:body] || ''
-      lexer = ops[:lexer] || ''
-      format = ops[:format] || ''
-
-      begin
-        require 'pygments'
-        begin
-          Pygments.highlight(
-                   unescape_html(body),
-                   :options => {
-                               :nowrap => true,
-                               :noclasses => true
-                             },
-                   :formatter => format,
-                   :lexer => lexer)
-        rescue MentosError
-          body
-        end
-      rescue LoadError
-          body
-      end
-    end
   end
 end   # module ReVIEW
